==> 2.download_image.ipynb <==
codis = pd.read_csv('data/codi.csv')
items = pd.read_csv('data/items.csv')

codis['img'].values

items['img'].values

!pwd

import os
import sys
import urllib 
import time

## 이미지 다운로드
def download_photo(img_url, filename, iscodi):

    if iscodi:
        file_path = "%s%s" % ("/Users/duyeoungryu/Documents/github/codi_recommendation/img/codis/", filename)
    else:
        file_path = "%s%s" % ("/Users/duyeoungryu/Documents/github/codi_recommendation/img/items/", filename)
    
    try:
        urllib.request.urlretrieve(img_url, file_path)
    except Exception as e:
        print(img_url, filename)
        print(e)
        
    return file_path

# call
#download_photo("http://img.mapssi.com/codi/vys6426@naver_com1521887849005.jpg", "test.jpg")

for idx in range(len(codis)):
    row = codis.iloc[idx]
    coid = row['id_codi']
    img_url = row['img']
    path = urllib.parse.urlparse(img_url).path
    ext = os.path.splitext(path)[1]
    
    newname = '%s%s' % (coid , ext)
    #print(coid,img_url, newname)
    
    download_photo(img_url, newname, True)

unique_id = list(set(items['id_item']))
for idx in unique_id:
    row = items[items['id_item'] == idx ].iloc[0]
    coid = row['id_item']
    img_url = row['img']
    path = urllib.parse.urlparse(img_url).path
    ext = os.path.splitext(path)[1]
    
    newname = '%s%s' % (coid , ext)
    #print(coid,img_url, newname)
    
    download_photo(img_url, newname, False)

len(codis.id_codi.unique())

!ls img/codis |wc -l

len(items.id_item.unique())

!ls img/items |wc -l
