==> 4.save_database.ipynb <==
%matplotlib inline
import re
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy import Column, Integer, String, DateTime
from sqlalchemy.orm import sessionmaker
import sqlalchemy
import pickle

from konlpy.tag import *
hannanum = Hannanum()
kkma = Kkma()
twitter = Twitter()

codis = pd.read_csv('data/codis.csv')
items = pd.read_csv('data/items.csv')

with open('pw.pw', 'rb') as f:
    pw = pickle.load(f) 

#ip = '13.125.171.94'
ip = '127.0.0.1'
db = 'codi'

engine = sqlalchemy.create_engine("mysql+mysqldb://root:"+pw+"@"+ip+"/"+db)

# 코디 저장

codis = codis.sort_values('rank').reset_index(drop=True)

codis.to_sql(name='codis', con=engine, if_exists='replace', index_label='id')

# 태그 분석

corpus = []
#분리하기
for idx in range(len(codis)):
    
    st = codis['tag'][idx]

    try:
        res = re.split('[#\n ]+',st)
    except Exception as e:
        res = ['']
        print(st)
        
    if '' in res:
        res.remove('')
    res = list(set(res))
    
    corpus.append(' '.join(res))

from sklearn.feature_extraction.text import CountVectorizer
 
vect = CountVectorizer()
vect.fit(corpus)

vc = vect.transform(corpus).toarray()
vc.shape

# 코디, 태그 매핑 정보 저장

rows = []
for idx in range(1485):
    tag1 = vc[idx]
    id_codi = codis.iloc[idx]['id_codi']
    idxs = np.argwhere(tag1==1).ravel().tolist()
    for cidx in idxs:
        rows.append([id_codi, cidx])

mapping_codi_tag = pd.DataFrame(rows, columns=['id_codi','id_tag'])

mapping_codi_tag.to_sql(name='mapping_codi_tag',  con=engine, if_exists='replace', index_label='id')

# 태그 저장

tagdf = pd.DataFrame(vect.get_feature_names(), columns=['tag'])

tagdf.to_sql(name="tags", con=engine, if_exists='replace', index_label='id')

len(items['id_item'].value_counts())

len(items)

# 아이템 저장

def get_price(x):
    arr = re.split('[→()]', x)
    
    res = '' 
    if len(arr) == 1:
        res = x
    else:
        # 할인정보 표시 지우기
        res = re.split('[→()]', x)[1]
    return res

items['price'] = items['price'].apply(get_price)
items['price'] = items['price'].astype(int)

# 코디와 아이템 매핑 정보
mapping_codi_item = items.pivot_table(index=['id_codi', 'id_item']).reset_index().drop(columns=['price'])

mapping_codi_item.to_sql(name="mapping_codi_item", con=engine, if_exists='replace', index_label='id')

# 중복 아이템 지우기
items = items.drop_duplicates(subset=['id_item']).reset_index(drop=True)

items.to_sql(name='items', con=engine, if_exists='replace', index_label='id')

# select * 
# from mapping_codi 
# join codis on mapping_codi.id_codi = codis.id_codi
# join tags on mapping_codi.id_tag = tags.id
# where codis.id_codi = '396015'

# select * 
# from `mapping_codi_item`
# join codis on `mapping_codi_item`.`id_codi` = `codis`.`id_codi`
# join items on `mapping_codi_item`.`id_item` = `items`.`id_item`
# where codis.id_codi= 395456
