==> 6.analysis_color.ipynb <==
%matplotlib inline

import pandas as pd
from ipywidgets import widgets

df_codis_info = pd.read_csv("data/codis_info.csv")

### RGB Color 확인(전체)

ar_codis_RGB = df_codis_info.filter(regex="_[RGB]").values

color_RGB = ar_codis_RGB.reshape(1116*3, 3) / 256
color_RGB.shape

def plot_color(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(color_RGB[:,0], color_RGB[:,1], color_RGB[:,2], c=color_RGB, s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)

widgets.interact(plot_color, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

### HSV Color 확인(전체)

ar_codis_HSV = df_codis_info.filter(regex="_[HSV]").values.reshape(1116*3, 3)

def plot_color_HSV(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(ar_codis_HSV[:,0], ar_codis_HSV[:,1], ar_codis_HSV[:,2], c=color_RGB, s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
widgets.interact(plot_color_HSV, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

- 무채색 계통이 많이 검출된다.

### Main Color1 (RGB, HSV)

ar_color1_RGB = df_codis_info.filter(regex="color1_[RGB]").values /256

def plot_color1(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(ar_color1_RGB[:,0], ar_color1_RGB[:,1], ar_color1_RGB[:,2], c=ar_color1_RGB, s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
    
widgets.interact(plot_color1, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

ar_color1_HSV = df_codis_info.filter(regex="color1_[HSV]").values

def plot_color1_HSV(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(ar_color1_HSV[:,0], ar_color1_HSV[:,1], ar_color1_HSV[:,2], c=ar_color1_RGB, s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
    
widgets.interact(plot_color1_HSV, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

- 검정과 흰색에 집중되어 있다.
- 중간이 거의 없다.

### Main Color2 (RGB, HSV)

ar_color2_RGB = df_codis_info.filter(regex="color2_[RGB]").values /256

def plot_color2(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(ar_color2_RGB[:,0], ar_color2_RGB[:,1], ar_color2_RGB[:,2], c=ar_color2_RGB, s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
    
widgets.interact(plot_color2, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

ar_color2_HSV = df_codis_info.filter(regex="color2_[HSV]").values

def plot_color2_HSV(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(ar_color2_HSV[:,0], ar_color2_HSV[:,1], ar_color2_HSV[:,2], c=ar_color2_RGB, s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
    
widgets.interact(plot_color2_HSV, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

### Main Color3 (RGB, HSV)

ar_color3_RGB = df_codis_info.filter(regex="color3_[RGB]").values /256

def plot_color3(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(ar_color2_RGB[:,0], ar_color3_RGB[:,1], ar_color3_RGB[:,2], c=ar_color3_RGB, s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
    
widgets.interact(plot_color3, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

ar_color3_HSV = df_codis_info.filter(regex="color3_[HSV]").values

def plot_color3_HSV(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(ar_color3_HSV[:,0], ar_color3_HSV[:,1], ar_color3_HSV[:,2], c=ar_color3_RGB, s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
    
widgets.interact(plot_color3_HSV, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

df_codis_info.tail()

### 전체 분포 확인 (RGB)

from sklearn.decomposition import PCA

ar_codis_RGB = df_codis_info.filter(regex="[RGB]").values
clt = PCA(3)
pca_RGB = clt.fit_transform(ar_codis_RGB)/256

from mpl_toolkits.mplot3d import Axes3D

def plot_pca(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(pca_RGB[:,0], pca_RGB[:,1], pca_RGB[:,2], c='b', s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
    
widgets.interact(plot_pca, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));

### 전체 분포 확인 (HSV)

ar_codis_HSV = df_codis_info.filter(regex="[HSV]").values
clt = PCA(3)
pca_HSV = clt.fit_transform(ar_codis_HSV)

pca_HSV[:,2]

def plot_pca_HSV(azim):
    fig = plt.figure(figsize = (10, 10))
    ax = fig.add_subplot(111, projection='3d');
    ax.scatter(pca_HSV[:,0], pca_HSV[:,1], pca_HSV[:,2], s=20, cmap=plt.cm.jet, alpha=1)
    ax.view_init(20, azim)
    
widgets.interact(plot_pca_HSV, azim=widgets.IntSlider(min=0,max=180,step=10,value=0));
